+++ myems-web/src/layouts/MyEMSRoutes.js
@@ -195,6 +195,7 @@ import DistributionSystem from '../components/MyEMS/AuxiliarySystem/Distribution
 // Microgrid
 import Microgrid from '../components/MyEMS/Microgrid/Microgrid';
 import MicrogridDetails from '../components/MyEMS/Microgrid/MicrogridDetails';
+import MicrogridReport from '../components/MyEMS/Microgrid/MicrogridReport';
 // FDD
 import FDDFault from '../components/MyEMS/FDD/Fault';
 // Monitoring
@@ -450,7 +451,8 @@ const MyEMSRoutes = () => (
 
     {/*Microgrid*/}
     <Route path="/microgrid" exact component={Microgrid} />
-    <Route path="/microgriddetails" exact component={MicrogridDetails} />
+    <Route path="/microgrid/details" exact component={MicrogridDetails} />
+    <Route path="/microgrid/report" exact component={MicrogridReport} />
 
     {/*FDD*/}
     <Route path="/fdd" exact component={FDDFault} />